requirements.txt
flask
gunicorn
line-bot-sdk==1.8.0
google-api-python-client
oauth2client
